# spec/openapi.yaml
openapi: 3.0.1
info:
  title: YumYum Delivery Microservice
  description: Delivery Microservice for YumYum
  version: 1.0.0
servers:
  - url: 'http://localhost:8081/v1'
...
paths:
...
  /vendor/isInRange:
    get:
      tags:
        - Vendor
      summary: Is customer in range from vendor
      description: Returns true if a given customer address is in range from the specified vendor according to the vendors delivery radius
      operationId: isCustomerInRangeFromVendor
      parameters:
        - name: userId
          in: header
          required: true
          schema:
            $ref: '#/components/schemas/id'
          example: 1015
        - name: vendorId
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/id'
          example: 158
        - name: customerLocation
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/Location'
        - name: vendorLocation
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/Location'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  isInRange:
                    type: boolean
                    example: true

        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                code: 400
                message: Bad Request
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                code: 403
                message: Forbidden
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                code: 404
                message: Not Found
...
  /vendor/{vendorId}/couriers:
    get:
      tags:
        - Vendor
        - Courier
      summary: Get vendor specific couriers
      description: Get all couriers that are hired by the provided vendor
      operationId: getCouriersByVendor
      parameters:
        - name: userId
          in: header
          required: true
          schema:
            $ref: '#/components/schemas/id'
          example: 1015
        - name: vendorId
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/id'
          example: 158
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PrivateCourier'
                example:
                  - courierId: 1816
                    vendorId: 158
                  - courierId: 1817
                    vendorId: 159
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                code: 400
                message: Bad Request
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                code: 403
                message: Forbidden
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                code: 404
                message: Not Found
...
components:
  schemas:
    id:
      type: integer
      description: Schema for our ids, keeps them all constant
      format: int64
    PrivateCourier:
      type: object
      x-class-extra-annotation: "@javax.persistence.Entity  @javax.persistence.Table(name = \"couriers\")"
      description: Courier that is privately hired by a vendor. Can only be assigned to the vendor that hired him
      properties:
        courierId:
          type: integer
          format: int64
          x-field-extra-annotation: "@javax.persistence.Id @javax.persistence.GeneratedValue(strategy = javax.persistence.GenerationType.IDENTITY)"
        vendorId:
          $ref: '#/components/schemas/id'
...
    Error:
      type: object
      x-class-extra-annotation: "@javax.persistence.Embeddable"
      x-field-extra-annotation: "@javax.persistence.Embedded"
      description: Error with error code and description
      properties:
        code:
          type: integer
        message:
          type: string
      example:
        code: 400
        message: Bad Request
...
    Location:
      type: object
      x-class-extra-annotation: "@javax.persistence.Embeddable"
      x-field-extra-annotation: "@javax.persistence.Embedded"
      description: Address
      properties:
        country:
          type: string
        city:
          type: string
        address:
          type: string
        postalCode:
          type: string
      example:
        country: Netherlands
        city: Delft
        address: Mekelweg 4
        postalCode: 2628 CD
...